parts_warehouse/requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.1.13
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.5.0
factory-boy==3.3.0
Faker==22.6.0
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pymongo==3.12.3
pytest==8.0.0
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.4
six==1.16.0
sqlparse==0.2.4
tzdata==2023.4
